@W-10120337: handler for spinnaker exception (#24)
* refactor(web/exceptions): create BaseExceptionHandlers class

to pave the way for an additional exception hander in kork-retrofi

@W-10120337

* feat(retrofit): add SpinnakerRetrofitExceptionHandlers and handle SpinnakerServerException and SpinnakerHttpException

SpinnakerRetrofitErrorHandler (introduced in https://github.com/spinnaker/kork/pull/821 /
https://git.soma.salesforce.com/spinnaker/kork/pull/13) converts RetrofitErrors into
SpinnakerServerException errors (or one of its subclasses), but there was no corresponding
exception handler for SpinnakerServerException.  So, the generic handleException method
from kork-web's GenericExceptionHandlers handled SpinnakerServerException and logged a
warning with a whole stack trace in addition to the error.

So, this commit adds a handler for SpinnakerServerException and SpinnakerHttpException to
restore the pre-SpinnakerRetrofitErrorHandler behavior.  GenericExceptionHandlers does
have a handler for RetrofitErrors, but SpinnakerServerException is now responsible for
retrieving information from the underlying RetrofitError and storing it in the exception's
message.

@W-10120337

* feat(retrofit): make SpinnakerHttpException, SpinnakerServerException chainable

* add license header