@W-11838491: backport upstream PRs to support the account api (#85)
* feat(credentials): easily create customizable credential repositories (#823)

@W-11838491

* feat(credentials): create repository one type at a time (#828)

* feat(credentials): easily create customizable credential repositories

* feat(credentials): misc improvements to credentials repository creation:
 - avoid double loading
 - support parallel loading

* fix(credentials/test): restore postconstruct, fix tests

* fix(credentials): clean code

@W-11838491

* feat(secrets): Add UserSecret API (#942)

This adds a new UserSecret API to kork-secrets. These secrets provide
Fiat role requirements in order to use these secrets. A basic
"opaque" user secret type is introduced for transporting Fiat role
requirements inside a secret payload, though other types of user
secrets can be introduced through an extension mechanism.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>

@W-11838491

* feat(kork-secrets): Add AccessControlled and user secrets updates (#971)

* feat(kork-secrets): Add AccessControlled and user secrets updates

- Introduces AccessControlled interface to allow fiat to delegate hasPermission() checks to the AccessControlled object itself.
- Introduces UserSecretType annotation for identifying user secret type discriminators for Jackson data-binding.
- Splits up UserSecret into delegate classes UserSecretData and UserSecretMetadata.
- Renames UserSecretMapper to UserSecretSerde to split into a strategy interface along with addition of UserSecretSerdeFactory.
- Adds a "string" user secret type for plain inline secrets like kubeconfig files.
- Updates SecretsManagerSecretEngine to use updated user secrets API.

* refactor(kork-secrets-aws): Use enum constant

* docs(kork-credentials-aws): Clarify credentials origin

* feat(kork-secrets-aws): Allow custom client factory

This adds a bean to customize what AWSSecretsManager client to use for a given map of secret parameters.

@W-11838491

* fix(kork-secrets-aws): use localstack 0.13.2

introduced by https://github.com/spinnaker/kork/pull/971

since 0.11.3 isn't available in SFCI.

@W-11838491

* fix(kork-secrets-aws): Fix default client bean configuration (#973)

This fixes a problem with the default configuration when used in other services.

@W-11838491

* Add some missing credentials and secrets APIs (#974)

* feat(credentials-api): Add CredentialsType annotation

This adds a type discriminator annotation for CredentialsDefinition
implementations which is useful for serialization and deserialization.

* feat(kork-secrets): Add UserSecretManager external secrets APIs

* update(secrets): log warning for invalid EncryptedSecret::tryParse

* test(secrets): add test for EncryptedSecret::tryParse

* feat(core): add NamedTypeParser strategy for ObjectMapperSubtypeConfigurer

This adds a way to customize the type discriminator discovery for an ObjectMapper. Usually, JsonTypeName is used from the Jackson library, though custom annotations can also be defined such as CredentialsType in kork-credentials-api.

@W-11838491

* feat(secrets): Add credentials masking in JSON (#992)

This adds a new `Sensitive` annotation to explicitly mark fields as
sensitive. A sensitive field will not be serialized to JSON unless its
value is a `secret://...` URI to avoid leaking credentials. This
serializer also provides a configurable regular expression to match
against `CredentialsDefinition` property names to mark them as likely
sensitive which applies masking in addition to logging a warning so that
the affected class can be updated.

@W-11838491

Co-authored-by: Nicolas Cohen <ncknt@users.noreply.github.com>
Co-authored-by: Matt Sicker <boards@gmail.com>
Co-authored-by: Matt Sicker <msicker@apple.com>