fix(retrofit): catch conversion errors in SpinnakerServerException constructor
The e.getBodyAs call in the SpinnakerServerException will throw a parsing exception if the retrofit response body does not match the expected json response type. This was resulting in RuntimeExceptions being throw during the construction of a SpinnakerServerException, which makes error tracing really hard. Reason being, we are expecting a SpinnakerServerException, but instead we see a RuntimeException.

This has bad downstream effects in the SpinnakerRetrofitExceptionHandlers class, as that class (which is supposed to handle SpinnakerServerExceptions coming from retrofit client requests) doesn't handle the error, and the handling instead takes place in the default controller advice.

Bugs during exception construction are sneaky...

@W-13132536